#266: Refactoring
