+++ analyzer/Makefile
@@ -29,17 +29,17 @@ IKXMLFILE=pgns-ik.xml
 IKJSONFILE=pgns-ik.json
 HEADERS=analyzer.h pgn.h lookup.h
 COMMONDIR=../common
-COMMON=$(COMMONDIR)/common.c $(COMMONDIR)/common.h $(COMMONDIR)/license.h
+COMMON=$(COMMONDIR)/common.c $(COMMONDIR)/parse.c $(COMMONDIR)/common.h $(COMMONDIR)/license.h
 
 CFLAGS?=-Wall -O2
 
 all: $(TARGETS)
 
 $(ANALYZER): analyzer.c pgn.c lookup.c $(HEADERS) $(COMMON) Makefile
-	$(CC) $(CPPFLAGS) $(CFLAGS) $(LDFLAGS) -o $(ANALYZER) -I$(COMMONDIR) pgn.c analyzer.c lookup.c $(COMMONDIR)/common.c $(LDLIBS$(LDLIBS-$(@)))
+	$(CC) $(CPPFLAGS) $(CFLAGS) $(LDFLAGS) -o $(ANALYZER) -I$(COMMONDIR) pgn.c analyzer.c lookup.c $(COMMONDIR)/common.c $(COMMONDIR)/parse.c $(LDLIBS$(LDLIBS-$(@)))
 
 $(ANALYZER_EXPLAIN): analyzer-explain.c pgn.c lookup.c $(HEADERS) $(COMMON) Makefile
-	$(CC) $(CPPFLAGS) $(CFLAGS) $(LDFLAGS) -o $(ANALYZER_EXPLAIN) -I$(COMMONDIR) pgn.c analyzer-explain.c lookup.c $(COMMONDIR)/common.c $(LDLIBS$(LDLIBS-$(@)))
+	$(CC) $(CPPFLAGS) $(CFLAGS) $(LDFLAGS) -o $(ANALYZER_EXPLAIN) -I$(COMMONDIR) pgn.c analyzer-explain.c lookup.c $(COMMONDIR)/common.c $(COMMONDIR)/parse.c $(LDLIBS$(LDLIBS-$(@)))
 
 json: $(ANALYZER_EXPLAIN) pgns2json.xslt
 	$(ANALYZER_EXPLAIN) -explain-xml >$(XMLFILE) && xsltproc pgns2json.xslt $(XMLFILE) >$(JSONFILE)
