+++ common/common.c
@@ -856,64 +856,3 @@ bool parseConst(const char **msg, const char *str)
   }
   return false;
 }
-
-bool parseFastFormat(StringBuffer *in, RawMessage *msg)
-{
-  unsigned int prio;
-  unsigned int pgn;
-  unsigned int src;
-  unsigned int dst;
-  unsigned int bytes;
-
-  char        *p;
-  int          i;
-  int          b;
-  unsigned int byt;
-  int          r;
-
-  p = strchr(sbGet(in), '\n');
-  if (!p)
-  {
-    return false;
-  }
-
-  // Skip the timestamp
-  p = strchr(sbGet(in), ',');
-  if (!p)
-  {
-    return false;
-  }
-
-  r = sscanf(p, ",%u,%u,%u,%u,%u,%n", &prio, &pgn, &src, &dst, &bytes, &i);
-  if (r == 5)
-  {
-    // now store the timestamp, unchanged
-    memset(msg->timestamp, 0, sizeof msg->timestamp);
-    memcpy(msg->timestamp, sbGet(in), CB_MAX(p - sbGet(in), sizeof msg->timestamp - 1));
-
-    msg->prio = prio;
-    msg->pgn  = pgn;
-    msg->src  = src;
-    msg->dst  = dst;
-    msg->len  = bytes;
-
-    p += i - 1;
-
-    for (b = 0; b < CB_MIN(bytes, FASTPACKET_MAX_SIZE); b++)
-    {
-      if ((sscanf(p, ",%x%n", &byt, &i) == 1) && (byt < 256))
-      {
-        msg->data[b] = byt;
-      }
-      else
-      {
-        logError("Unable to parse incoming message '%s' data byte %u\n", sbGet(in), b);
-        return false;
-      }
-      p += i;
-    }
-    return true;
-  }
-  logError("Unable to parse incoming message '%s', r = %d\n", sbGet(in), r);
-  return false;
-}
