+++ ikonvert-serial/Makefile
@@ -25,8 +25,8 @@ CFLAGS= -Wall -O2
 
 all: $(TARGETS)
 
-$(IKONVERT): ikonvert-serial.c ikonvert.h ../common/common.c ../common/b64.c ../common/common.h
-	$(CC) $(CPPFLAGS) $(CFLAGS) $(LDFLAGS) -o $(IKONVERT) -I../common ikonvert-serial.c ../common/common.c ../common/b64.c $(LDLIBS$(LDLIBS-$(@)))
+$(IKONVERT): ikonvert-serial.c ikonvert.h ../common/common.c ../common/parse.c ../common/b64.c ../common/common.h
+	$(CC) $(CPPFLAGS) $(CFLAGS) $(LDFLAGS) -o $(IKONVERT) -I../common ikonvert-serial.c ../common/common.c ../common/parse.c ../common/b64.c $(LDLIBS$(LDLIBS-$(@)))
 
 clean:
 	-rm -f $(TARGETS) *.elf *.gdb
